Public/WEB/SystemIntegration_STIs/src/requirements.txt
mysql-connector-python==9.0.0
flask==2.3.3